third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/worker-in-sandboxed-iframe-by-csp-fetch-event.https-expected.txt

This is a testharness.js-based test.
PASS Prepare a service worker.
FAIL Prepare an iframe sandboxed by CSP HTTP header with allow-scripts. assert_false: Service worker should NOT control the sandboxed page expected false got true
PASS Prepare an iframe sandboxed by CSP HTTP header with allow-scripts and allow-same-origin.
FAIL Fetch request from a worker in iframe sandboxed by CSP HTTP header allow-scripts flag assert_equals: The request should NOT be handled by SW. expected 0 but got 1
PASS Fetch request from a worker in iframe sandboxed by CSP HTTP header with allow-scripts and allow-same-origin flag
Harness: the test ran to completion.
